requirements.txt
langchain==0.0.283
llama-cpp-python==0.1.83
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
BingImageCreator==0.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
EdgeGPT==0.12.0
frozenlist==1.3.3
GoogleBard==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hugchat==0.0.8
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.8
prompt-toolkit==3.0.39
Pygments==2.15.1
python-dateutil==2.8.2
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
semaphore-bot==0.16.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tqdm==4.65.0
urllib3==2.0.3
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
